Move files to tensorflow module
